Boundary curve function now generalised for most basic inlet shapes
A small value of alpha greater than 0 seems to correctly get the nodes at the boundary of most simple shapes. A value of 0 gives the convex hull which ignores the nodes at the sides of rectangles (outputting only the corner points).
The alpha parameter is hard coded as 0.1, for more complex inlet shapes this may need to be editable in the future. A call to alphashape.optimizealpha would be needed, but this takes a long time.
Also don't need the conditional cellSides and numCells calculation in init