Added capability for translating contour plots to a flow field
Class added for reading in images containing the contour plots and colour bars. Needs both the contour plots of the tangential flow angle and the radial flow angle to fully reconstruct the velocity field.  Also needs the ranges of the colour bars, the duct radius and the mean axial velocity.